% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all_ireland}
\alias{all_ireland}
\title{Generalised NUTS3 regions for the island of Ireland (including Northern Ireland)}
\description{
A \link[sp]{SpatialPolygonsDataFrame} of NUTS3 regions in the island of Ireland,  with some associated variables
}
\details{
Data frame contains the following variables;
\describe{
 \item{NUTS_ID}{NUTS3 Region code}
 \item{STAT_LEVL_}{Statistical level}
 \item{SHAPE_AREA}{Region area}
 \item{NUTS2NAME}{Region perimeter}
}
}

