% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\docType{package}
\name{caricRture}
\alias{caricRture}
\alias{caricRture-package}
\title{A tool set for creating 'stylised' maps and other graphics.}
\description{
The key tools here are functions to create 'sketchy' polygons, lines and points,  and
'curvey' polygons and lines.  Used in combination they can create grap]hics having a 'hand drawn'
appearance in a number of styles - such as pencil-drawn or marker pen. The 'sketchy' 
renderings are based on the algorithms of \href{https://web.archive.org/web/20140913142528/http://openaccess.city.ac.uk/1274/}{Wood et. al.}.
These are useful in conveying ther fact that some data-driven graphics are based 
on vague or fuzzy information.
}

