% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{choro_sketch_it}
\alias{choro_sketch_it}
\title{Draw a sketchy choropleth map}
\usage{
choro_sketch_it(sp, v, shading = auto.shading, ...)
}
\arguments{
\item{sp}{A  \link[sp]{SpatialPolygonsDataFrame} object}

\item{v}{A variable (or expression) to control the colouring of the map}

\item{shading}{A shading scheme - see \link[GISTools]{auto.shading}}

\item{...}{other parameters passed to \link[GISTools]{choropleth}}
}
\value{
The input \code{sp} - useful for pipelining
}
\description{
Draw a sketchy choropleth map
}
\examples{
x <- 7

}

