% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clone_data}
\alias{clone_data}
\title{Clone data from a \link[sp]{SpatialPolygonsDataFrame}}
\usage{
clone_data(sp,spdf)
sp \%>\% clone_data(spdf)
}
\arguments{
\item{sp}{a \link[sp]{SpatialPolygons} object}

\item{spdf}{a \link[sp]{SpatialPolygonsDataFrame} object with matching IDs}
}
\value{
a \link[sp]{SpatialPolygonsDataFrame}  object with polygons from \code{sp} and data from \code{spdf}.
}
\description{
Populate a transformed \link[sp]{SpatialPolygons} object with data from a
\link[sp]{SpatialPolygonsDataFrame} with an identical set of IDs
}
\examples{
data(RA)
RA.spdf \%>\% small_chop -> RA.sp
class(RA.sp)
RA.sp \%>\% clone_data(RA.spdf) -> RA.sp2
RA.sp2 \%>\% data.frame \%>\% head
}

