% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{contract_it}
\alias{contract_it}
\title{Contract polygons in an object}
\usage{
contract_it(spdf, r)
}
\arguments{
\item{spdf}{A \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}

\item{r}{Radius of contraction}
}
\value{
A \link[sp]{SpatialPolygons} object with contraction applied
}
\description{
Contract polygons in an object
}
\examples{
"POLYGON ((30 10, 40 40, 20 40, 10 20, 30 10))" \%>\% readWKT -> p1
dv <- adjustcolor('darkviolet',alpha.f=0.5)
p1 \%>\% dilate_it(3) \%>\% plot(col=dv)
p1 \%>\% plot(col=dv,add=TRUE)

# A sketchy version
p1 \%>\% contract_it(3) \%>\% make_canvas \%>\% sketch_it(col='chocolate')
}

