% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{curve_it}
\alias{curve_it}
\title{'Curvify' polygon-based objects.}
\usage{
curve_it(spdf,s)
spdf %>% curve_it(s)
}
\arguments{
\item{s}{\code{shape} parameter as in \link[graphics]{xspline}}

\item{spdf}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
a \link[sp]{SpatialPolygons} curved caricature
}
\description{
Curved caracture from a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame}
object, controlled by a shape parameter.  This can pass through the nodes of the
original object (-1 < shape parameter < 0) or go near to them (0 < shape parameter < 1).
}
\details{
This is based on the \code{\link[graphics]{xspline}} function. In particular,
the shape parameter is the same as in that function.
}
\examples{
'indianred' \%>\% adjustcolor(alpha.f=0.3) -> ired
"POLYGON((0 0,0 2,1 3.5,3 3,4 1,3 0,0 0))" \%>\% readWKT -> p1
p1 \%>\% plot(col=ired)
p1 \%>\% curve_it(1) \%>\% plot(add=TRUE,col=ired,lty=2)
p1 \%>\% curve_it(0.5) \%>\% plot(add=TRUE,col=ired,lty=2)
}

