% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{dilate_it}
\alias{dilate_it}
\title{Dilate (expand) polygons in an object}
\usage{
dilate_it(spdf,r)
spdf %>% dilate_it(r)
}
\arguments{
\item{spdf}{A \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}

\item{r}{Radius of dilation}
}
\value{
A \link[sp]{SpatialPolygons} object with the dilation applied
}
\description{
Dilate the polygons in a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object
}
\examples{
"POLYGON ((30 10, 40 40, 20 40, 10 20, 30 10))" \%>\% readWKT -> p1
dsg <- adjustcolor('darkseagreen',alpha.f=0.5)
p1 \%>\% dilate_it(3) \%>\% plot(col=dsg)
p1 \%>\% plot(col=dsg,add=TRUE)

# A sketchy version
p1 \%>\% dilate_it(3) \%>\% make_canvas \%>\% sketch_it(col='navyblue')
}

