% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{fuse_data}
\alias{fuse_data}
\title{Fuse data from to \link[sp]{Spatial} object}
\usage{
fuse_data(sp,v)
sp \%>\% fuse_data(v)
}
\arguments{
\item{sp}{a \link[sp]{SpatialPolygons} object}

\item{v}{a variables whose names or rownames match thev IDs for \code{sp}}
}
\value{
a \link[sp]{SpatialPolygonsDataFrame}  object with polygons from \code{sp} and data from \code{v}.
}
\description{
Populate a transformed \link[sp]{SpatialPolygons} object with data from a 
matrix or array with a set of names or rownames matching the \link[sp]{SpatialPolygons} object IDs
}
\examples{
data(RA)
RA.spdf \%>\% small_chop -> RA.sp
class(RA.sp)
RA.sp \%>\% clone_data(RA.spdf) -> RA.sp2
RA.sp2 \%>\% data.frame \%>\% head
}

