% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{get_fonts}
\alias{get_fonts}
\title{Install some Google handwriting-style fonts}
\usage{
get_fonts()
}
\value{
Nothing is returned
}
\description{
Use this function once to install some hand-written style fonts
(from Google Fonts \url{https://www.google.com/fonts}) into R.
}
\details{
For this function to run successfully, computer must be connected to the internet.
At the time of writing,  \pkg{caricRture} uses \pkg{showtext} to handle the fonts.  That package doesn't
work well with the RStudio graphics device - better to use \code{quartz} or \code{X11} to write to a floating
window if using RStudio. Each font has a two-letter abbreviation,  which can be used as a \code{family} parameter
in \code{par}.  The abbreviations and associated fonts are:
\describe{
\item{am}{Amatic SC}
\item{dk}{Dekko}
\item{gr}{Covered By Your Grace}
\item{hl}{Handlee}
\item{pm}{Permanent Marker}
\item{rs}{Rock Salt}
\item{ws}{Waiting for the Sunrise}
}
}
\examples{
get_fonts() # Thats basically it - once run,  fonts will be installed
}

