% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{hand_compass}
\alias{hand_compass}
\title{Sketchy compass with 'handwritten' font}
\usage{
hand_compass <- function (x, y, rot = 0, cex = 1, names =  c("E", "N", "W", "S"), overdraw=FALSE, ...) 
x \%>\% hand_compass(...)
}
\arguments{
\item{x}{x-coordinate for compass centre}

\item{y}{y-coordinate for compass centre}

\item{r}{compass 'radius' in map units}

\item{rot}{compass rotation}

\item{cex}{compass scale}

\item{names}{Names for the four compass directions}

\item{overdraw}{whether to use overdraw when drawing compass lines}

\item{...}{other parameters passed on to the \link{sketch_it} function}

\item{fontfam}{Handwritten font family - see \link{get_fonts}}
}
\value{
No value returned
}
\description{
Adds a compass (very crude) to a hand drawn map.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied 
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
get_fonts()
data(RA) 
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% clone_data(RA.spdf) -> RA.spdf2
shadecol <- ifelse(RA.spdf2$NUTS2 == 'IE01','indianred','dodgerblue')

RA.spdf2 \%>\% make_canvas \%>\% sketch_it(col=shadecol)
c('NUTS2 I','NUTS2 II') \%>\% hand_legend(-22450,427691,45000,15000,col=c('indianred','dodgerblue'),cex=1.8)

}

