% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{hand_legend}
\alias{hand_legend}
\title{Sketchy legend with 'handwritten' font}
\usage{
hand_legend <- function(labels,tlx,tly,side,gap,cex=1,fontfam='am',...)
labels \%>\% hand_legend(...)
}
\arguments{
\item{labels}{a character string array of legends labels}

\item{tlx}{top left x coordinate of legend}

\item{tly}{top left y coordinate of legend}

\item{side}{length of size of square in legend}

\item{gap}{gap between bottom of one legend and top of next}

\item{cex}{scale of legend text}

\item{fontfam}{Handwritten font family - see \link{get_fonts}}

\item{...}{other parameters passed on to the \link{sketch_it} function}
}
\value{
No value returned
}
\description{
Similar to the \link[graphics]{legend} command,  
but chooses a hand writing styled font,  and draws sketchy legend
symbols
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied 
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
get_fonts()
data(RA) 
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% clone_data(RA.spdf) -> RA.spdf2
shadecol <- ifelse(RA.spdf2$NUTS2 == 'IE01','indianred','dodgerblue')

RA.spdf2 \%>\% make_canvas \%>\% sketch_it(col=shadecol)
c('NUTS2 I','NUTS2 II') \%>\% hand_legend(-22450,427691,45000,15000,col=c('indianred','dodgerblue'),cex=1.8)

}

