% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{hand_text}
\alias{hand_text}
\title{Text with 'handwritten' font}
\usage{
hand_text(txt,x,y,fontfam='am',,...)
txt \%>\% hand_text(x,y,fontfam='am',...)
}
\arguments{
\item{x}{x coordinate of text location}

\item{y}{y coordinate of text location - if missing, loks in \code{x} for both coordinates}

\item{fontfam}{Handwritten font family - see \link{get_fonts}}

\item{...}{other parameters passed on to the \link[graphics]{text} function}

\item{txt}{a character string to write as the title}
}
\value{
No value returned
}
\description{
Works like the \link[graphics]{text} command,  but chooses a hand writing styled font.
Also,  parameter order is changed so that the text label comes first - useful for pipelining with \code{\%>\%}.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied 
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
get_fonts()
data(RA) 
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% make_canvas \%>\% sketch_it(col='lightgrey',lwd=0.5)
RA.spdf \%>\% coordinates -> label_points
RA.spdf$NUTS3NAME \%>\% hand_text(label_points,col='darkred',fontfam='pm',cex=0.7)
}

