% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hand_title}
\alias{hand_title}
\title{Title with 'handwritten' font}
\usage{
hand_title(txt,fontfam='am',...)
txt \%>\% hand_title(...)
}
\arguments{
\item{fontfam}{Handwritten font family - see \link{get_fonts}}

\item{...}{other parameters passed on to the \link[graphics]{title} function}

\item{txt}{a character string to write as the title}
}
\value{
No value returned
}
\description{
Works like the \link[graphics]{title} command,  but chooses a hand writing styled font.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
get_fonts()
data(RA)
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% make_canvas \%>\% sketch_it(col='orange')
'NUTS3 Regions in Ireland' \%>\% hand_title(fontfam='gr')
}

