% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{hole_chop}
\alias{hole_chop}
\title{Remove internal holes in polygons}
\usage{
hole_chop(spdf)
spdf \%>\% hole_chop
}
\arguments{
\item{spdf}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
a \link[sp]{SpatialPolygons}  object with internal holes removed (IDs are the same as \code{spdf})
}
\description{
Remove internal holes in a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object leaving topolgy intact
}
\examples{
data(RA)
RA.spdf \%>\% hole_chop \%>\% plot
}

