% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{Hpdf}
\alias{Hpdf}
\title{Convenience driver for PDFs with hand written font}
\usage{
Hpdf(file)
file \%>\% Hpdf(...)
}
\arguments{
\item{...}{other parameters passed on to the \link[grDevices]{pdf} function}

\item{file}{file name to write to}
}
\value{
No value returned
}
\description{

}
\details{
Works like the \link[grDevices]{pdf} command,  but specifies a hand writing styled font 'Bradley Hand Bold'.

You'll need to have the 'Bradley Hand Bold' font installed
}
\examples{
# Create a sketchy map and write to a pdf file
data(RA)
'test.pdf' \%>\% Hpdf(h=6,w=10)
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% make_canvas \%>\% sketch_it(col='lightgrey',lwd=0.5)
RA.spdf \%>\% coordinates -> label_points
RA.spdf$NUTS3NAME \%>\% Htext(label_points,col='darkred')
dev.off()
}

