% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{Htext}
\alias{Htext}
\title{Text with 'handwritten' font}
\usage{
Htext(txt,x,y,...)
txt \%>\% Htitle(x,y,...)
}
\arguments{
\item{x}{x coordinate of text location}

\item{y}{y coordinate of text location}

\item{...}{other parameters passed on to the \link[graphics]{title} function}

\item{txt}{a character string to write as the title}
}
\value{
No value returned
}
\description{
Works like the \link[graphics]{text} command,  but chooses a hand writing styled font 'Bradley Hand Bold'.
Also,  parameter order is changed so that the text label comes first - useful for pipelining with \code{\%>\%}.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
data(RA)
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% make_canvas \%>\% sketch_it(col='lightgrey',lwd=0.5)
RA.spdf \%>\% coordinates -> label_points
RA.spdf$NUTS3NAME \%>\% Htext(label_points,col='darkred')
}

