% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{Htitle}
\alias{Htitle}
\title{Title with 'handwritten' font}
\usage{
Htitle(txt,...)
txt \%>\% Htitle(...)
}
\arguments{
\item{...}{other parameters passed on to the \link[graphics]{title} function}

\item{txt}{a character string to write as the title}
}
\value{
No value returned
}
\description{
Works like the \link[graphics]{title} command,  but chooses a hand writing styled font 'Bradley Hand Bold'.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
data(RA)
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% make_canvas \%>\% sketch_it(col='orange')
'NUTS3 Regions in Ireland' \%>\% Htitle
}

