% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hull_it}
\alias{hull_it}
\title{Make all polygons in an object convex}
\usage{
hull_it(spdf)
spdf %>% hull_it
}
\arguments{
\item{spdf}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
a \link[sp]{SpatialPolygons} object with convex hulls of polygons (may overlap)
}
\description{
Replace polygons in \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} with convex hulls
}
\examples{
"MULTIPOLYGON (((30 20, 45 40, 35 35, 25 35, 10 40, 30 20)),
   ((15 5, 40 8, 40 10, 35 8, 10 20, 5 10, 15 5)))" \%>\% readWKT -> p1
'lavender' \%>\% adjustcolor(alpha.f=0.5) \%>\% lav
p1 \%>\% plot(col=lav)
p1 \%>\% hull_it \%>\% plot(col=lav,add=TRUE)
}

