% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{londonhp}
\alias{londonhp}
\title{House price data for Greater London}
\description{
A \link[sp]{SpatialPolygonsDataFrame} of London Bouroughs (\code{londonborough}) and a
\link[sp]{SpatialPointsDataFrame} of London house price data (\code{londonhp}).  The house price data is for 2001
and was supplied by the Nationwide Building Society.  There are 316 house price sales and 20
variables.
}
\details{
The variables are:
\describe{
 \item{PURCHASE}{Purchase price (UKP)}
 \item{FLOORSZ}{Floor size (square metres)}
 \item{TYPEDETCH,TPSEMIDTCH,TYPETRRD,TYPEBNGLW,TYPEFLAT}{House type dummy: Detached, Semi, Terraced,Bungalow,Flat}
 \item{BLDPWW1,BLDINTW,BLDPOSTW,BLD60S,BLD70S,BLD80S,BLD90S}{Dummy for period of building:Pre WW1, Inter-War, Post WW2, or recent decade}
 \item{BATH2}{Two or more bathrooms - indicator}
 \item{BEDS2}{Two bedrooms - indicator}
 \item{GARAGE1}{Garage present - indicator}
 \item{UNEMPLOY}{Unemployment rate in census ward of property}
 \item{PROF}{Rate of employees in professional or managerial employment in census ward of property}
}
}

