% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{make_canvas}
\alias{make_canvas}
\title{Create a new canvas}
\usage{
make_canvas(spdf)
spdf \%>\% make_canvas
}
\arguments{
\item{spdf}{A \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
Echos \code{spdf} - this is useful for pipelines using \code{\%>\%}
}
\description{
Create a new plot whose extent fits a given \link[sp]{Spatial} object. Returns the input spatial object,
which is very useful for pipelines.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied 
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
data(RA) 
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% make_canvas \%>\% sketch_it(col='orange')
}

