% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{outline_it}
\alias{outline_it}
\title{Find the outline of a group of polygons}
\usage{
outline_it(spdf)
spdf %>% outline_it
}
\arguments{
\item{spdf}{A \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame}}
}
\value{
A \link[sp]{SpatialPolygons} outline of the input object
}
\description{
Outline from a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object
- any internal holes are filled in.
}
\examples{
data(RA)
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% outline_it \%>\% make_canvas \%>\% plot_it
}

