% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{sketch_it}
\alias{sketch_it}
\title{Sketchy drawing of a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
\usage{
sketch_it(spdf,...)
spdf \%>\% sketch_it(...)
}
\arguments{
\item{...}{parameters passed to \code{\link[graphics]{lines}}}

\item{spdf}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\description{
Draws a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object in a 'hand drawn'
style.
}
\examples{
"POLYGON ((30 10, 40 40, 20 40, 10 20, 30 10))" \%>\% readWKT -> p1
# Create a blank canvas with extent containing p1
p1 \%>\% make_canvas \%>\% sketch_it(col='indianred')
}

