% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{small_chop}
\alias{small_chop}
\title{Cut out small islands and enclaves}
\usage{
small_chop(spdf)
spdf \%>\% small_chop
}
\arguments{
\item{spdf}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
a \link[sp]{SpatialPolygons}  object with satellite polygons removed (IDs are the same as \code{spdf})
}
\description{
Remove satellite polygons in a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object.
For each compound polygon,  remove all but the largest component.  Useful for generalised representations.
}
\examples{
data(RA)
RA.spdf \%>\% small_chop \%>\% plot
'MULTIPOLYGON (((3 2, 4.5 4, 1 4, 3 2)),((15 5, 40 10, 10 20, 5 10, 15 5)))' \%>\% readWKT -> p1
'darksalmon' \%>\% adjustcolor(alpha.f=0.5) -> salmon
old.mf <- par('mfrow')
par(mfrow=c(1,2))
p1 \%>\% plot(col=salmon); title("Before")
p1 \%>\% small_chop \%>\% plot(col=salmon); title("After")
par(mfrow=old.mf)
}

