% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/caricRture.R
\name{tidy_it}
\alias{tidy_it}
\title{Tidy up an object with overlapping polygons}
\usage{
tidy_it(x)
}
\arguments{
\item{x}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
a \link[sp]{SpatialPolygons} object with overlaps cut out
}
\description{
Cut out overlaps in a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object
}
\examples{
# Create an object with two overlapping polygons
"MULTIPOLYGON (((30 20, 45 40, 10 40, 30 20)))" \%>\% readWKT(id="George") -> george
"MULTIPOLYGON (((15 5, 40 10, 41 35, 10 20, 5 10, 15 5)))" \%>\% readWKT(id="Mildred") -> mildred
rbind(george,mildred) -> p1
c("George","Mildred") -> ID

# Create a color scheme based on IDs of polygons in p1
'olivedrab' \%>\% adjustcolor(alpha.f=0.5) -> olive
'lemonchiffon' \%>\%  adjustcolor(alpha.f=0.5) -> lemon
c(George=olive,Mildred=lemon) -> cammo

# Useful for the demo
function(x) sapply(slot(x,'polygons'), function(y) slot(y,'ID')) -> get.IDs

par('mfrow') -> old.mf
par(mfrow=c(1,2))
p1 \%>\% plot(col=cammo[p1 \%>\% get.IDs]); 'Before' \%>\% title
p1 \%>\% tidy_it -> p1_tidy
p1_tidy \%>\% plot(col=cammo[p1_tidy \%>\% get.IDs]); 'After' \%>\% title
par(mfrow=old.mf)
}

