% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caricRture.R
\name{hand_compass}
\alias{hand_compass}
\title{Sketchy compass with 'handwritten' font}
\usage{
hand_compass <- function (x, y, r, rot = 0, cex = 1, north_name =  "N", fontfam='am', overdraw=FALSE, ...) 
x \%>\% hand_compass(...)
}
\arguments{
\item{x}{x-coordinate for compass centre}

\item{y}{y-coordinate for compass centre}

\item{r}{compass 'radius' in map units}

\item{rot}{compass rotation}

\item{cex}{compass scale}

\item{north_name}{Name for north to write on the compass}

\item{fontfam}{Handwritten font family - see \link{get_fonts}}

\item{overdraw}{whether to use overdraw when drawing compass lines}

\item{...}{other parameters passed on to the \link{sketch_it} function}
}
\value{
No value returned
}
\description{
Adds a compass (very crude) to a hand drawn map.
}
\examples{
# Here make_canvas is inserted in a pipeline after a simplified and tidied 
# map of Irish NUTS3 regions is created.  A sketchy rendition is then added to the
# canvas.
get_fonts()
data(RA) 
RA.spdf \%>\% small_chop \%>\% gSimplify(tol=11000) \%>\% tidy_it \%>\% clone_data(RA.spdf) -> RA.spdf2
shadecol <- ifelse(RA.spdf2$NUTS2 == 'IE01','indianred','dodgerblue')

RA.spdf2 \%>\% make_canvas \%>\% sketch_it(col=shadecol)
# Add a north arrow with 'North' written in Irish
hand_compass(353685,424980,30000,fontfam = 'pm',north_name="ó thuaidh")

}

