% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sketch_it}
\alias{sketch_it}
\title{Sketchy drawing of a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
\usage{
sketch_it(spdf,rough=0.05,...)
spdf \%>\% sketch_it(rough=0.05,...)
}
\arguments{
\item{rough}{controls the 'roughness' of the sketched polygon edges}

\item{...}{parameters passed to \code{\link[graphics]{lines}}}

\item{spdf}{a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object}
}
\value{
the input \code{spdf} - useful for pipelines
}
\description{
Draws a \link[sp]{SpatialPolygons} or \link[sp]{SpatialPolygonsDataFrame} object in a 'hand drawn'
style.
}
\examples{
"POLYGON ((30 10, 40 40, 20 40, 10 20, 30 10))" \%>\% readWKT -> p1
# Create a blank canvas with extent containing p1
p1 \%>\% make_canvas \%>\% sketch_it(col='indianred')
}

